client/boonhook/requirements.txt
flask
gevent
PyJWT>=2.0
boonsdk
